fix(app): 官方 provider 常量移到无依赖叶子模块,修复 renderer crypto 崩溃
app 从 core/provider import OFFICIAL_PROVIDER_IDS/isOfficialProvider 会经
schema→util/hash 拖入 node crypto(createHash),Vite 为浏览器 externalize
后 renderer 一加载即崩。把这三个纯常量抽到无 import 的 provider-official.ts,
app 改从该叶子模块 import;provider.ts re-export 保持后端不变。
