+++ packages/app/src/hooks/use-providers.ts
@@ -3,7 +3,7 @@ import { decode64 } from "@/utils/base64"
 import { useParams } from "@solidjs/router"
 import { Iterable, pipe } from "effect"
 import { createMemo } from "solid-js"
-import { OFFICIAL_PROVIDER_IDS } from "@deepagent-code/core/provider"
+import { OFFICIAL_PROVIDER_IDS } from "@deepagent-code/core/provider-official"
 
 // Official providers are the recommended-first display set and the only ids the backend treats as
 // first-party (key store credentials, fixed protocol). Everything else is a custom third-party
